requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
APScheduler==3.10.4
httpx==0.25.2
datasets==3.6.0
huggingface-hub==0.30.1
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.4
transformers>=4.30.0
torch
fastapi
uvicorn
python-multipart
pillow 